+++ .env_example
@@ -7,7 +7,7 @@ RECAPTCHA_PUBLIC_KEY=REPLACE_WITH_KEY
 RECAPTCHA_SECRET_KEY=REPLACE_WITH_KEY
 KUE_USER=user
 KUE_PASS=pass
-DELETE_REPLAYS=1
+DELETE_REPLAYS=false
 AWS_S3_BUCKET=somebucket
 AWS_ACCESS_KEY_ID=REPLACE_WITH_KEY
 AWS_SECRET_ACCESS_KEY=REPLACE_WITH_KEY
@@ -19,3 +19,4 @@ REDIS_URL=redis://127.0.0.1:6379
 STEAM_USER=user1,user2
 STEAM_PASS=pass1,pass2
 STEAM_GUARD_CODE=""
+RETRIEVER=false
